+++ R/to_miniature.R
@@ -29,11 +29,6 @@
 #' images, see [image_border()].
 #' @param fileout if not NULL, result is saved in a png file whose filename
 #' is defined by this argument.
-#' @param use_docx2pdf if TRUE (and if 'Microsoft Word' executable
-#' can be found as well as 'docx2pdf'), docx2pdf will be used to
-#' convert 'Word' documents to PDF. This makes it possible to have a
-#' PDF identical to the 'Word' display whereas with 'LibreOffice', this
-#' is not always the case.
 #' @param timeout timeout in seconds that libreoffice is allowed to use
 #' in order to generate the corresponding pdf file, ignored if 0.
 #' @return a magick image object as returned by [image_read()].
@@ -43,8 +38,8 @@
 #'   package = "doconv",
 #'   "doc-examples/example.docx"
 #' )
-#' if(exec_available("python") && docx2pdf_available())
-#'   to_miniature(docx_file, use_docx2pdf = TRUE)
+#' if(exec_available("word"))
+#'   to_miniature(docx_file)
 #'
 #' pptx_file <- system.file(
 #'   package = "doconv",
@@ -52,9 +47,10 @@
 #' )
 #' if(exec_available("libreoffice") && check_libreoffice_export())
 #'   to_miniature(pptx_file)
+#' @importFrom locatexec pip_exec exec_available
 to_miniature <- function(filename, row = NULL, width = NULL,
                          border_color = "#ccc", border_geometry = "2x2",
-                         fileout = NULL, use_docx2pdf = FALSE, timeout = 120) {
+                         fileout = NULL, timeout = 120) {
 
   if (!file.exists(filename)) {
     stop("filename does not exist")
@@ -65,15 +61,13 @@ to_miniature <- function(filename, row = NULL, width = NULL,
     pptx_to_miniature(
       filename, row = row, width = width,
       border_color = border_color, border_geometry = border_geometry,
-      fileout = fileout,
-      timeout = timeout)
+      fileout = fileout, timeout = timeout)
   } else if(grepl("\\.(doc|docx)$", filename)){
     if(is.null(width)) width <- 650
     docx_to_miniature(
       filename, row = row, width = width,
       border_color = border_color, border_geometry = border_geometry,
-      fileout = fileout, use_docx2pdf = use_docx2pdf,
-      timeout = timeout)
+      fileout = fileout, timeout = timeout)
   } else if(grepl("\\.pdf$", filename)){
     if(is.null(width)) width <- 650
     pdf_to_miniature(
@@ -92,7 +86,7 @@ pdf_to_miniature <- function(filename, row = NULL, width = 650,
   img_list <- pdf_to_images(filename)
   x <- images_to_miniature(
     img_list = img_list,
-    row = row, width = width,
+    row = row, width = width * 150/72,
     border_color = border_color, border_geometry = border_geometry
   )
   if(!is.null(fileout))
@@ -102,13 +96,14 @@ pdf_to_miniature <- function(filename, row = NULL, width = 650,
 
 docx_to_miniature <- function(filename, row = NULL, width = 650,
                               border_color = "#ccc", border_geometry = "2x2",
-                              fileout = NULL, use_docx2pdf = FALSE,
-                              timeout = 120) {
+                              fileout = NULL, timeout = 120) {
   pdf_filename <- tempfile(fileext = ".pdf")
 
-  if(use_docx2pdf && exec_available("word") && docx2pdf_available())
+  if (exec_available("word")) {
     docx2pdf(input = filename, output = pdf_filename)
-  else to_pdf(input = filename, output = pdf_filename, timeout = timeout)
+  } else {
+    to_pdf(input = filename, output = pdf_filename, timeout = timeout)
+  }
 
   x <- pdf_to_miniature(pdf_filename,
     row = row, width = width,
@@ -121,10 +116,14 @@ docx_to_miniature <- function(filename, row = NULL, width = 650,
 
 pptx_to_miniature <- function(filename, row = NULL, width = 750,
                               border_color = "#ccc", border_geometry = "2x2",
-                              fileout = NULL,
-                              timeout = 120) {
+                              fileout = NULL, timeout = 120) {
   pdf_filename <- tempfile(fileext = ".pdf")
-  to_pdf(input = filename, output = pdf_filename, timeout = timeout)
+  if (exec_available("powerpoint")) {
+    pptx2pdf(input = filename, output = pdf_filename)
+  } else {
+    to_pdf(input = filename, output = pdf_filename, timeout = timeout)
+  }
+
   x <- pdf_to_miniature(pdf_filename,
     row = row, width = width,
     border_color = border_color, border_geometry = border_geometry