+++ R/to_pdf.R
@@ -10,11 +10,6 @@
 #' @param input,output file input and optional file output. If output
 #' file is not provided, the value will be the value of input file with
 #' extension "pdf".
-#' @param use_docx2pdf if TRUE (and if 'Microsoft Word' executable
-#' can be found as well as 'docx2pdf'), docx2pdf will be used to
-#' convert 'Word' documents to PDF. This makes it possible to have a
-#' PDF identical to the 'Word' display whereas with 'LibreOffice', this
-#' is not always the case.
 #' @param UserInstallation use this value to set a non-default user profile path
 #' for "LibreOffice". If not provided a temporary dir is created. It makes possibles
 #' to use more than a single session of "LibreOffice."
@@ -48,7 +43,7 @@
 #' @importFrom locatexec libreoffice_exec
 #' @importFrom locatexec is_windows
 to_pdf <- function(input, output = gsub("\\.[[:alnum:]]+$", ".pdf", input),
-                   use_docx2pdf = FALSE, timeout = 120, UserInstallation = NULL) {
+                   timeout = 120, UserInstallation = NULL) {
   if (!file.exists(input)) {
     stop("input does not exist")
   }
@@ -74,36 +69,35 @@ to_pdf <- function(input, output = gsub("\\.[[:alnum:]]+$", ".pdf", input),
   }
 
 
-  if(grepl("\\.(doc|docx)$", input) && use_docx2pdf && exec_available("word") && docx2pdf_available()){
+  if(grepl("\\.(doc|docx)$", input) && exec_available("word")){
     docx2pdf(input = input, output = output)
+  } else if(grepl("\\.(ppt|pptx)$", input) && exec_available("powerpoint")){
+    pptx2pdf(input = input, output = output)
   } else {
     file_set_origin <- list.files(default_root, full.names = TRUE)
-    suppressWarnings(
-      info <- try(
-        system2(
+
+    res <- run(
           libreoffice_exec(),
           args = c("--headless",
-                   if(!is_windows()) sprintf("\"-env:UserInstallation=file://%s\"", UserInstallation)
-                   else sprintf("\"-env:UserInstallation=file:///%s\"", UserInstallation),
+                   if(!is_windows()) sprintf("-env:UserInstallation=file://%s", UserInstallation)
+                   else sprintf("-env:UserInstallation=file:///%s", UserInstallation),
                    "--convert-to", "pdf:writer_pdf_Export",
-                   "--outdir", shQuote(default_root, type = "cmd"),
-                   shQuote(input, type = "cmd")),
-          stderr = TRUE, stdout = TRUE, timeout = timeout), silent = TRUE)
-    )
-    out <- !1 %in% attr(info, "status")
-    if(!out) {
-      stop(paste0(info, collapse = "\n"))
-    }
+                   "--outdir", default_root,
+                   input),
+          error_on_status = FALSE)
+    success <- res$status == 0
+
+    if(!success) stop("could not convert ", input, call. = FALSE)
+
     file_set_new <- list.files(default_root, full.names = TRUE)
     file_set_new <- setdiff(file_set_new, file_set_origin)
     success <- file.copy(from = file_set_new, to = output, overwrite = TRUE)
     unlink(file.path(default_root, basename(input)), force = TRUE)
     unlink(file_set_new, force = TRUE)
     if (any(!success)) {
-      stop("could not convert ", shQuote(input))
+      stop("could not convert ", input, call. = FALSE)
     }
   }
-
 
   invisible(output)
 }
@@ -134,7 +128,7 @@ check_libreoffice_export <- function(UserInstallation = NULL) {
 
   file.copy(
     system.file(package = "doconv", "doc-examples", "minimal-word.docx"),
-    default_root
+    default_root, overwrite = TRUE
   )
 
   if(is.null(UserInstallation)){
@@ -144,21 +138,20 @@ check_libreoffice_export <- function(UserInstallation = NULL) {
     UserInstallation <- absolute_path(UserInstallation)
   }
 
-  suppressWarnings(
-    try(
-      system2(
-        libreoffice_exec(),
-        args = c("--headless",# useless unless with older versions
-                 if(!is_windows()) sprintf("\"-env:UserInstallation=file://%s\"", UserInstallation)
-                 else sprintf("\"-env:UserInstallation=file:///%s\"", UserInstallation),
-                 "--convert-to", "pdf:writer_pdf_Export",
-                 "--outdir", shQuote(default_root, type = "cmd"),
-                 shQuote(input, type = "cmd")),
-        stderr = TRUE, stdout = TRUE, timeout = 15), silent = TRUE)
-  )
+  res <- run(command = libreoffice_exec(),
+      args = c("--headless",# useless unless with older versions
+               if(!is_windows()) sprintf("-env:UserInstallation=file://%s", UserInstallation)
+               else sprintf("-env:UserInstallation=file:///%s", UserInstallation),
+               "--convert-to", "pdf:writer_pdf_Export",
+               "--outdir", default_root,
+               input),
+      error_on_status = FALSE,
+      timeout = 15
+      )
+  success <- res$status == 0
+
   expected <- file.path(default_root, "minimal-word.pdf")
   success <- file.exists(expected)
 
   success
 }
-